provisioner/requirements.txt
requests>=2.31.0
Jinja2>=3.1.4
PyYAML>=6.0.1
